Check open connections in unix before deleting. Candidate for #5

+++ R/misc-functions.R
@@ -9,7 +9,7 @@
   
   # Parse file name and path
   if (file == "") file <- paste0("ukb-schemas-", date_str, ".sqlite")
-  full_path <- paste0(path.expand(path), "\\", file)
+  full_path <- paste0(path.expand(path), "/", file)
   
   # If `file`` exists, session is `interactive()` and `!overwrite`, then prompt 
   # to overwrite the file
@@ -22,8 +22,18 @@
     }
     if (!isTRUE(overwrite)) 
       stop(UKBSCHEMAS_ERRORS$OVERWRITE)
-    else tryCatch(
-      file.remove(full_path),
+    else tryCatch({
+      if (.Platform$OS.type == "unix"){
+          sys_command <- paste("lsof", full_path, "| wc -l")
+          processes_using_file <- as.numeric(system(sys_command, intern = TRUE,
+                                                    ignore.stdout = TRUE,
+                                                    ignore.stderr = TRUE))
+          if (processes_using_file>1) {
+            stop(UKBSCHEMAS_ERRORS$FAILED_OVERWRITE)
+          }
+        }
+        file.remove(full_path)
+      },
       error = function(err) {
         stop(UKBSCHEMAS_ERRORS$FAILED_OVERWRITE)
       },
